+++ src/packages/eslint-plugin-relay-fragments/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+- possibility to pass `threshold` for `limit-complexity` ruled added, e.g. `"limit-complexity": ["error", { "threshold": 30 }]`
+
 # 0.0.1
 
 - `limit-complexity` rule added